+++ cf/field.py
@@ -7148,9 +7148,9 @@ class Field(mixin.FieldDomain, mixin.PropertiesData, cfdm.Field):
             )
             domain_axis = collapse_axes.get(healpix_axis)
             if domain_axis is not None and domain_axis.get_size() > 1:
-                from .healpix_utils import _del_healpix_coordinate_reference
+                from .healpix_utils import del_healpix_coordinate_reference
 
-                _del_healpix_coordinate_reference(f)
+                del_healpix_coordinate_reference(f)
 
             # ---------------------------------------------------------
             # Update dimension coordinates, auxiliary coordinates,