+++ Changelog.rst
@@ -11,7 +11,7 @@ Version NEXTVERSION
   (https://github.com/NCAS-CMS/cf-python/issues/???)
 * New method: `cf.Data.coarsen`
   (https://github.com/NCAS-CMS/cf-python/issues/???)
-* New optional dependency: ``healpix>=2024.2``
+* New optional dependency: ``healpix>=2025.1``
 * Changed dependency: ``cfdm>=1.13.0.0, <1.13.1.0``
 
 ----